+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0](https://github.com/LippertComponents/Orchestrator/compare/v1.3.2...v1.4.0) - 2019-07-24
+### Added
+ - Deploy command will now require MODX extras as defined in the defined in the `core/config/lci_modx_transport_package.php` file
+ - New commands to manage MODX extras
+
+See all extras in your MODX install:
+``` 
+cd www/core/
+php vendor/bin/orchestrator extras 
+```
+
+Example on how to require an extra:
+```
+cd www/core/
+ php vendor/bin/orchestrator modx:package ace-1.8.0-pl 
+```
+ 
+
 ## [1.3.2](https://github.com/LippertComponents/Orchestrator/compare/v1.3.1...v1.3.2) - 2019-06-27
 ### Changed
  - Forgot to version bump LCI\MODX\Orchestrator\Console\Application